split low level xdr decoding from high
